Bump dependencies (#957)
