+++ .eslintrc.security
@@ -8,6 +8,7 @@
   "plugins": [
     "@typescript-eslint/tslint"
   ],
+  "ignorePatterns": ["**/__mocks__/**"],
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "project": "tsconfig.json",
