+++ __tests__/Auth0Client/buildLogoutUrl.test.ts
@@ -4,6 +4,7 @@ import { verify } from '../../src/jwt';
 import { MessageChannel } from 'worker_threads';
 import * as utils from '../../src/utils';
 import * as scope from '../../src/scope';
+import { expect } from '@jest/globals';
 
 // @ts-ignore
