+++ __tests__/Auth0Client/checkSession.test.ts
@@ -5,6 +5,7 @@ import { verify } from '../../src/jwt';
 import { MessageChannel } from 'worker_threads';
 import * as utils from '../../src/utils';
 import * as scope from '../../src/scope';
+import { expect } from '@jest/globals';
 
 // @ts-ignore
