+++ __tests__/Auth0Client/handleRedirectCallback.test.ts
@@ -5,6 +5,7 @@ import { MessageChannel } from 'worker_threads';
 import * as utils from '../../src/utils';
 import * as scope from '../../src/scope';
 import * as http from '../../src/http';
+import { expect } from '@jest/globals';
 
 // @ts-ignore
