+++ __tests__/helpers.ts
@@ -1,3 +1,5 @@
+import { expect } from '@jest/globals';
+
 export const expectToHaveBeenCalledWithAuth0ClientParam = (mock, expected) => {
   const [[url]] = (<jest.Mock>mock).mock.calls;
   const param = new URL(url).searchParams.get('auth0Client');
