+++ __tests__/utils.test.ts
@@ -16,6 +16,7 @@ import {
 } from '../src/utils';
 
 import { DEFAULT_AUTHORIZE_TIMEOUT_IN_SECONDS } from '../src/constants';
+import { expect } from '@jest/globals';
 
 (<any>global).TextEncoder = TextEncoder;
 
@@ -143,7 +144,7 @@ describe('utils', () => {
         subtle: {
           digest: jest.fn((alg, encoded) => {
             expect(alg).toMatchObject({ name: 'SHA-256' });
-            expect(Array.from(encoded)).toMatchObject([116, 101, 115, 116]);
+            expect(Array.from(encoded)).toEqual([116, 101, 115, 116]);
             return new Promise(res => res(true));
           })
         }
@@ -368,7 +369,7 @@ describe('utils', () => {
       expect(message.source.close).toHaveBeenCalled();
       expect(window.document.body.appendChild).toHaveBeenCalledWith(iframe);
       expect(window.document.body.removeChild).toHaveBeenCalledWith(iframe);
-      expect(iframe.setAttribute.mock.calls).toMatchObject([
+      expect(iframe.setAttribute.mock.calls).toEqual([
         ['width', '0'],
         ['height', '0'],
         ['src', url]
