+++ package.json
@@ -10,7 +10,7 @@
   "scripts": {
     "dev": "rimraf dist && rollup -c --watch",
     "start": "npm run dev",
-    "docs": "typedoc --options ./typedoc.js --module commonjs",
+    "docs": "typedoc --options ./typedoc.js src",
     "build": "rimraf dist && rollup -m -c --environment NODE_ENV:production && npm run test:es-check",
     "build:stats": "npm run build -- --environment WITH_STATS:true && open stats.html",
     "lint:security": "eslint ./src --ext ts --no-eslintrc --config ./.eslintrc.security",
@@ -27,36 +27,38 @@
     "test:integration": "concurrently --raw --kill-others --success first npm:test:integration:server npm:test:integration:tests",
     "serve:coverage": "serve coverage/lcov-report -n",
     "serve:stats": "serve bundle-stats -n",
-    "print-bundle-size": "node ./scripts/print-bundle-size",
+    "print-bundle-size": "node ./scripts/print-bundle-size.mjs",
     "prepack": "npm run build && node ./scripts/prepack",
     "publish:cdn": "ccu --trace"
   },
   "devDependencies": {
     "@auth0/component-cdn-uploader": "github:auth0/component-cdn-uploader#v2.2.2",
-    "@rollup/plugin-replace": "^2.4.2",
+    "@rollup/plugin-replace": "^4.0.0",
     "@types/cypress": "^1.1.3",
-    "@types/jest": "^27.0.2",
-    "@typescript-eslint/eslint-plugin-tslint": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
+    "@types/jest": "^28.1.7",
+    "@typescript-eslint/eslint-plugin-tslint": "^5.33.1",
+    "@typescript-eslint/parser": "^5.33.1",
     "browserstack-cypress-cli": "1.8.1",
     "cli-table": "^0.3.6",
-    "concurrently": "^6.4.0",
+    "concurrently": "^7.3.0",
     "cypress": "7.2.0",
-    "es-check": "^6.1.1",
-    "eslint": "^7.32.0",
-    "gzip-size": "^6.0.0",
+    "es-check": "^7.0.0",
+    "eslint": "^8.22.0",
+    "gzip-size": "^7.0.0",
     "husky": "^7.0.4",
     "idtoken-verifier": "^2.2.2",
-    "jest": "^27.5.1",
-    "jest-junit": "^13.0.0",
-    "jest-localstorage-mock": "^2.4.18",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "jest-junit": "^14.0.0",
+    "jest-localstorage-mock": "^2.4.22",
     "jsonwebtoken": "^8.5.1",
-    "oidc-provider": "^7.10.1",
-    "prettier": "^2.4.1",
+    "node-fetch": "^3.2.10",
+    "oidc-provider": "^7.11.5",
+    "prettier": "^2.7.1",
     "pretty-quick": "^3.1.2",
     "qss": "^2.0.3",
     "rimraf": "^3.0.2",
-    "rollup": "^2.60.0",
+    "rollup": "^2.78.0",
     "rollup-plugin-analyzer": "^4.0.0",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-dev": "^1.1.3",
@@ -64,16 +66,16 @@
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-sourcemaps": "^0.6.3",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "rollup-plugin-visualizer": "^5.5.2",
+    "rollup-plugin-typescript2": "^0.32.1",
+    "rollup-plugin-visualizer": "^5.7.1",
     "rollup-plugin-web-worker-loader": "^1.6.1",
-    "serve": "^12.0.1",
-    "ts-jest": "^27.0.7",
+    "serve": "^14.0.1",
+    "ts-jest": "^28.0.8",
     "tslib": "^2.4.0",
     "tslint": "^6.1.3",
     "tslint-config-security": "^1.16.0",
-    "typedoc": "0.18.0",
-    "typescript": "^4.4.4",
+    "typedoc": "^0.23.10",
+    "typescript": "^4.7.4",
     "wait-on": "^6.0.0"
   },
   "dependencies": {
