+++ typedoc.js
@@ -22,11 +22,7 @@ module.exports = {
     'src/worker/**/*',
     ...excludeFiles.map(f => `./src/${f}.ts`)
   ],
-  mode: 'file',
-  excludeExternals: true,
+  excludeExternals: false,
   excludePrivate: true,
-  includeDeclarations: true,
-  stripInternal: true,
-  hideGenerator: true,
-  theme: 'minimal'
+  hideGenerator: true
 };
